+++ src/main/java/net/sf/mpxj/LocaleData.java
@@ -2456,6 +2456,8 @@ public final class LocaleData extends ListResourceBundle
       PROJECT_COLUMNS_ARRAY[ProjectField.MINUTES_PER_DAY.getValue()] = "Minutes per Day";
       PROJECT_COLUMNS_ARRAY[ProjectField.DAYS_PER_MONTH.getValue()] = "Days per Month";
       PROJECT_COLUMNS_ARRAY[ProjectField.MINUTES_PER_WEEK.getValue()] = "Minutes per Week";
+      PROJECT_COLUMNS_ARRAY[ProjectField.MINUTES_PER_MONTH.getValue()] = "Minutes per Month";
+      PROJECT_COLUMNS_ARRAY[ProjectField.MINUTES_PER_YEAR.getValue()] = "Minutes per Year";
       PROJECT_COLUMNS_ARRAY[ProjectField.FISCAL_YEAR_START.getValue()] = "Fiscal Year Start";
       PROJECT_COLUMNS_ARRAY[ProjectField.DEFAULT_TASK_EARNED_VALUE_METHOD.getValue()] = "Default Tssk Earned Value Method";
       PROJECT_COLUMNS_ARRAY[ProjectField.REMOVE_FILE_PROPERTIES.getValue()] = "Remove File Properties";