+++ src/main/java/net/sf/mpxj/primavera/PrimaveraDatabaseReader.java
@@ -215,7 +215,9 @@ public final class PrimaveraDatabaseReader extends AbstractProjectReader
          ph.setCreationDate(row.getDate("create_date"));
          ph.setLastSaved(row.getDate("update_date"));
          ph.setMinutesPerDay(Integer.valueOf((int)(row.getDouble("day_hr_cnt").doubleValue() * 60)));
-         ph.setMinutesPerWeek(Integer.valueOf((int)(row.getDouble("week_hr_cnt").doubleValue() * 60)));
+         ph.setMinutesPerWeek(Integer.valueOf((int)(row.getDouble("week_hr_cnt").doubleValue() * 60)));         
+         ph.setMinutesPerMonth(Integer.valueOf((int)(row.getDouble("month_hr_cnt").doubleValue() * 60)));
+         ph.setMinutesPerYear(Integer.valueOf((int)(row.getDouble("year_hr_cnt").doubleValue() * 60)));
          ph.setWeekStartDay(Day.getInstance(row.getInt("week_start_day_num")));
 
          processDefaultCurrency(row.getInteger("curr_id"));